pyproject.toml
[project]
name = "oracle-rag"
version = "0.1.0"
description = "A powerful PDF RAG (Retrieval-Augmented Generation) system built with LangChain, designed to be wrapped as an MCP server"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "langchain-chroma>=0.1.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.1",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.1",
    "mypy>=1.11.1",
    "pytest>=8.0.0",
]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "UP",   # pyupgrade
]
lint.ignore = [
    "T201", # allow print()
    "E501",  # ignore line length
    "D101",  # ignore docstring for private modules
    "D103",  # ignore docstring for private modules
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
